// resources/src/assets/ts/model/entity/device_entity.ts
export type DeviceEntity = {
    uuid: string;
    type: string;
    device_name: string;
    device_username: string;
    device_password: string;
    device_host: string;
    device_mac: string;
    authorized: boolean;
    created_at: string;
    uptime: string;
    login: boolean;
    now_value: string;
}